# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CIBOL · 实验室系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/requests": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Create Request",
        "operationId": "create_request_api_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/requests/mine": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "My Requests",
        "description": "与我相关的全部请求（我发起的 + 收到的对调），按时间倒序。",
        "operationId": "my_requests_api_requests_mine_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "type": "array",
                  "title": "Response My Requests Api Requests Mine Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/requests/pending": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Pending Requests",
        "description": "审批中心队列：请假/API/SSH 中待管理员处理（submitted）的请求。",
        "operationId": "pending_requests_api_requests_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "type": "array",
                  "title": "Response Pending Requests Api Requests Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/requests/processed": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Processed Requests",
        "description": "审批中心历史：请假/API/SSH 中已通过或已驳回的，按时间倒序。",
        "operationId": "processed_requests_api_requests_processed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "type": "array",
                  "title": "Response Processed Requests Api Requests Processed Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/requests/{req_id}/advance": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Advance Request",
        "operationId": "advance_request_api_requests__req_id__advance_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "req_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Req Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvanceRequest": {
        "properties": {
          "next": {
            "type": "string",
            "enum": [
              "accepted",
              "declined",
              "cancelled",
              "approved",
              "rejected"
            ],
            "title": "Next"
          },
          "note": {
            "type": "string",
            "maxLength": 512,
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "next"
        ],
        "title": "AdvanceRequest"
      },
...
      "CreateRequest": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/RequestKind"
          },
          "fromDate": {
            "type": "string",
            "maxLength": 32,
            "title": "Fromdate",
            "default": ""
          },
          "toName": {
            "type": "string",
            "maxLength": 64,
            "title": "Toname",
            "default": ""
          },
          "toDate": {
            "type": "string",
            "maxLength": 32,
            "title": "Todate",
            "default": ""
          },
          "topic": {
            "type": "string",
            "maxLength": 256,
            "title": "Topic",
            "default": ""
          },
          "detail": {
            "type": "string",
            "maxLength": 256,
            "title": "Detail",
            "default": ""
          },
          "reason": {
            "type": "string",
            "maxLength": 4096,
            "title": "Reason",
            "default": ""
          },
          "note": {
            "type": "string",
            "maxLength": 512,
            "title": "Note",
            "default": ""
          },
          "fromMeetingId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frommeetingid"
          },
          "toMeetingId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tomeetingid"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "CreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequestEventOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "at"
        ],
        "title": "RequestEventOut"
      },
      "RequestKind": {
        "type": "string",
        "enum": [
          "swap",
          "absence",
          "api",
          "ssh"
        ],
        "title": "RequestKind"
      },
      "RequestOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/RequestKind"
          },
          "from": {
            "type": "string",
            "title": "From"
          },
          "toName": {
            "type": "string",
            "title": "Toname",
            "default": ""
          },
          "fromDate": {
            "type": "string",
            "title": "Fromdate",
            "default": ""
          },
          "toDate": {
            "type": "string",
            "title": "Todate",
            "default": ""
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "default": ""
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "incoming": {
            "type": "boolean",
            "title": "Incoming",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/RequestEventOut"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "from",
          "status",
          "createdAt",
          "history"
        ],
        "title": "RequestOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}